requirements.txt
# requirements file
# For Python 3.11.5

accelerate==0.30.1
ipython==8.24.0
ipywidgets==8.1.2
numpy==1.26.4
Pillow==10.3.0
quanto==0.1.0
torch==2.2.2
transformers==4.41.0
sentencepiece==0.2.0
seaborn==0.13.0
matplotlib==3.8.2
